Rename files to pip 8 standards.
