+++ vision/tools/blob_annotator/requirements.txt
@@ -1,1 +1,3 @@
 lxml
+opencv-python
+numpy
