+++ vision/tools/blob_annotator/tools/generate_annotation.py
