+++ vision/tools/blob_annotator/ui/Window.py
@@ -1,3 +1,6 @@
+"""
+ui window.
+"""
 import cv2
